MsgPack deserializer: improve handling of EOF
`rmp_serde` has two kinds of errors that contain I/O errors, and an
EOF can occur inside either of them, but we were only treating an EOF
inside an `InvalidMarkerRead` as an EOF, which would make sense for
the beginning of a message.

However, we should also treat an incomplete message + EOF as an EOF.
There isn't really any point in reporting that an EOF was received
mid-message.

This should fix the issue where the
`seq_describe_no_collect_succeeds_without_error` test would sometimes
fail, as doing a `describe --no-collect` followed by nushell exiting
could (but was not guaranteed to) cause this exact scenario.