Fix circular source causing Nushell to crash
EngineState now tracks the script currently running, instead of the
parent directory of the script. This also provides an easy way to expose
the current running script to the user (Issue #12195).

Similarly, StateWorkingSet now tracks scripts instead of directories.
`parsed_module_files`, which was used to prevent circular module import,
is renamed to `scripts` and acts like a stack for tracking the current
running script (which is on the top of the stack).

Circular import check is now extended for source as well. A simple
testcase is added for circular source.