Refactor `Record` to use a single backing `Vec`
This shrinks `Record`'s size in half and should make it again amenable
to be part of `Value` directly.

Might have slightly different performance characteristics as the cache
locality changes on lookups (if you directly need the value, it should
be closer, but in other cases may blow up the cache line budget)

Also different perf characteristics on creation expected.
`from_raw_cols_vals` now probably worse.

Comparison with the main branch (boxed Record) revealed no significant
change to the creation but an improvement when accessing larger N.
The fact that this was more pronounced for nested (still cloning before
nushell/nushell#12325) leads to the conclusion that this may still be
dominated by the smaller clone necessary for a 24-byte `Record` over the
previous 48 bytes.