Elide clone in `V::follow_cell_path` for record
This clone is not necessary and tanks the performance of deep nested
access.
As soon as we found the value we now we discard the old value so can
`std::mem::take` the inner (`impl Default for Value` to the rescue)

We may be able to further optimize this but not having to clone the
value is vital.