Make default config conservative about clipboard
Some platforms don't support the `system-clipboard` feature, notably
termux on android. The default config currently contained references to
`reedline` events that are only available with the feature enabled
(#12179). This thus broke the out of the box config for those users. For
now be more defensive about this and only enable default events. Add the
alternative as commented out code you can quickly enable.