Isolate tests from user config
This is an attempt to isolate the unit tests from whatever might be in
the user's config. If the user's config is broken in some way or
incompatible with this version (for example, especially if there are
plugins that aren't built for this version), tests can spuriously fail.

This makes tests more reliably pass the same way they would on CI even
if the user has config, and should also make them run faster.