Add extra support from the engine for foregrounding
New engine calls: `EnterForeground`, `ExitForeground`

This plays nicely with the way nushell creates a new process group and
hands terminal control to externals when they're run. On Unix,
`EnterForeground` will return a process group ID to join if the plugin
needs to join an existing group, or it will give terminal control to the
plugin if not.