Avoid allocating in `parse_datetime`
Make the preparsing stricter before trying
`chrono::DateTime::parse_from_rfc3339` which expects a fully qualified
combination of date, time, and timezone info.

If given the correct string now zero-copy. For just a date or date and
time without timezone we have to allocate to add the assumed midnight
and/or UTC.

Added prechecks may duplicate some work of the chrono parser but should
preclude allocations and full parses for wildly wrong strings.