Fix nu-check
nu-check handles modules and scripts differently. In case of modules, it calls parse_module_file which adds the file to the stack, but in case of scripts it doesn't. The current solution is to just add the file to the stack manually. It's not pretty, but it'll do for now.